chore: change project configuration
