+++ .prettierrc.cjs
@@ -1,8 +1,6 @@
-const config = {
+module.exports = {
   trailingComma: 'es5',
   tabWidth: 2,
-  semi: false,
+  semi: true,
   singleQuote: true,
 };
-
-module.exports = config;
