--- rss_react/src/App.tsx
@@ -1,5 +1,0 @@
-function App() {
-  return <div>Hello, React!</div>
-}
-
-export default App
