+++ src/App.tsx
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <div>
+      <h1>Vite + React, Hello!</h1>
+    </div>
+  );
+}
+
+export default App;
