+++ src/main.tsx
@@ -1,0 +1,11 @@
+/* eslint-disable @typescript-eslint/comma-dangle */
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+
+ReactDOM.createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
